feat: rename CreateTransactionsFromCrawledDataCommand file
- Added a new command class `CreateTransactionsFromCrawledDataCommand` to handle the crawling of receipt URLs and the creation of transaction records.
- The command validates crawled data, resolves shop addresses, and inserts transactions into the database.
- Updated the `__all__` exports in the crawl_url module to include the new command class, ensuring proper module organization.
